Update requirements
The package currently requires mlflow, which pulls both mlflow but also
many other packages which adds up to many gigs of data. Instead, depend
on mlflow-skinny, which is just the package with the mlflow bits, and
users can pick and choose whatever packages to add on top. There is an
additional dep on jinja to render the templates